Moved and renamed getConfig function
Changed terminology in useContainer descripition to be more accurate
