+++ src/shared/codelens/localLambdaRunner.ts
@@ -6,7 +6,6 @@
 import * as path from 'path'
 import * as tcpPortUsed from 'tcp-port-used'
 import * as vscode from 'vscode'
-import { getLocalLambdaConfiguration } from '../../lambda/local/configureLocalLambda'
 import { detectLocalLambdas, LocalLambda } from '../../lambda/local/detectLocalLambdas'
 import { CloudFormation } from '../cloudformation/cloudformation'
 import { makeTemporaryToolkitFolder } from '../filesystemUtilities'
@@ -22,7 +21,7 @@ import { SamTemplateGenerator } from '../templates/sam/samTemplateGenerator'
 import { ExtensionDisposableFiles } from '../utilities/disposableFiles'
 
 import { writeFile } from 'fs-extra'
-import { generateDefaultHandlerConfig, HandlerConfig } from '../../lambda/config/templates'
+import { getHandlerConfig, HandlerConfig } from '../../lambda/config/templates'
 import { DebugConfiguration } from '../../lambda/local/debugConfiguration'
 import { getFamily, RuntimeFamily } from '../../lambda/models/samLambdaRuntime'
 import { getLogger, Logger } from '../logger'
@@ -91,7 +90,7 @@ export class LocalLambdaRunner {
 
             const inputTemplate: string = await this.generateInputTemplate(this.codeRootDirectoryPath)
 
-            const config = await getConfig({
+            const config = await getHandlerConfig({
                 handlerName: this.localInvokeParams.handlerName,
                 documentUri: this.localInvokeParams.document.uri,
                 samTemplate: this.localInvokeParams.samTemplate
@@ -184,7 +183,7 @@ export class LocalLambdaRunner {
 
         const eventPath: string = path.join(await this.getBaseBuildFolder(), 'event.json')
         const environmentVariablePath = path.join(await this.getBaseBuildFolder(), 'env-vars.json')
-        const config = await getConfig({
+        const config = await getHandlerConfig({
             handlerName: this.localInvokeParams.handlerName,
             documentUri: this.localInvokeParams.document.uri,
             samTemplate: this.localInvokeParams.samTemplate
@@ -401,7 +400,7 @@ export async function invokeLambdaFunction(
 
     const eventPath: string = path.join(invokeArgs.baseBuildDir, 'event.json')
     const environmentVariablePath = path.join(invokeArgs.baseBuildDir, 'env-vars.json')
-    const config = await getConfig({
+    const config = await getHandlerConfig({
         handlerName: invokeArgs.originalHandlerName,
         documentUri: invokeArgs.documentUri,
         samTemplate: vscode.Uri.file(invokeArgs.originalSamTemplatePath)
@@ -458,25 +457,6 @@ export async function invokeLambdaFunction(
             })
         }
     }
-}
-
-export async function getConfig(params: {
-    handlerName: string
-    documentUri: vscode.Uri
-    samTemplate: vscode.Uri
-}): Promise<HandlerConfig> {
-    const workspaceFolder = vscode.workspace.getWorkspaceFolder(params.documentUri)
-    if (!workspaceFolder) {
-        return generateDefaultHandlerConfig()
-    }
-
-    const config: HandlerConfig = await getLocalLambdaConfiguration(
-        workspaceFolder,
-        params.handlerName,
-        params.samTemplate
-    )
-
-    return config
 }
 
 const getEnvironmentVariables = (
